Inject UiThreadExecutor into delegates
