+++ thirtyinch/src/main/java/net/grandcentrix/thirtyinch/TiActivity.java
@@ -22,6 +22,7 @@ import net.grandcentrix.thirtyinch.internal.TiActivityDelegate;
 import net.grandcentrix.thirtyinch.internal.TiLoggingTagProvider;
 import net.grandcentrix.thirtyinch.internal.TiPresenterProvider;
 import net.grandcentrix.thirtyinch.internal.TiViewProvider;
+import net.grandcentrix.thirtyinch.internal.UiThreadExecutor;
 import net.grandcentrix.thirtyinch.util.AndroidDeveloperOptions;
 import net.grandcentrix.thirtyinch.util.AnnotationUtil;
 
@@ -32,6 +33,7 @@ import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
 
 import java.util.List;
+import java.util.concurrent.Executor;
 
 /**
  * Created by pascalwelsch on 9/8/15.
@@ -90,6 +92,11 @@ public abstract class TiActivity<P extends TiPresenter<V>, V extends TiView>
         return null;
     }
 
+    @Override
+    public Executor getUiThreadExecutor() {
+        return new UiThreadExecutor();
+    }
+
     /**
      * Invalidates the cache of the latest bound view. Forces the next binding of the view to run
      * through all the interceptors (again).
@@ -134,11 +141,6 @@ public abstract class TiActivity<P extends TiPresenter<V>, V extends TiView>
         }
 
         return null;
-    }
-
-    @Override
-    public boolean postToMessageQueue(final Runnable runnable) {
-        return getWindow().getDecorView().post(runnable);
     }
 
     @SuppressWarnings("unchecked")
