+++ thirtyinch/src/main/java/net/grandcentrix/thirtyinch/internal/UiThreadExecutorAutoBinder.java
@@ -18,10 +18,6 @@ package net.grandcentrix.thirtyinch.internal;
 import net.grandcentrix.thirtyinch.TiLifecycleObserver;
 import net.grandcentrix.thirtyinch.TiPresenter;
 
-import android.os.Handler;
-import android.os.Looper;
-import android.support.annotation.NonNull;
-
 import java.util.concurrent.Executor;
 
 /**
@@ -29,34 +25,14 @@ import java.util.concurrent.Executor;
  */
 public class UiThreadExecutorAutoBinder implements TiLifecycleObserver {
 
-    /**
-     * Executes work on the UI thread. If the current thread is the UI thread, then the action is
-     * executed immediately. If the current thread is not the UI thread, the action is posted to the
-     * event queue of the UI thread.
-     */
-    private static class UiThreadExecutor implements Executor {
-
-        private final Handler mHandler = new Handler(Looper.getMainLooper());
-
-        private Thread mUiThread = Looper.getMainLooper().getThread();
-
-        @Override
-        public void execute(@NonNull Runnable command) {
-            if (Thread.currentThread() == mUiThread) {
-                // already on main thread, simply execute
-                command.run();
-            } else {
-                mHandler.post(command);
-            }
-        }
-    }
-
     private final TiPresenter mPresenter;
 
-    private final UiThreadExecutor mUiThreadExecutor = new UiThreadExecutor();
+    private final Executor mUiThreadExecutor;
 
-    public UiThreadExecutorAutoBinder(final TiPresenter presenter) {
+    public UiThreadExecutorAutoBinder(final TiPresenter presenter,
+            final Executor uiThreadExecutor) {
         mPresenter = presenter;
+        mUiThreadExecutor = uiThreadExecutor;
     }
 
     @Override
